# use default java serialization again