--- a/kernel/pom.xml
+++ b/kernel/pom.xml
@@ -181,12 +181,6 @@
       <classifier>windows</classifier>
       <type>exe</type>
     </dependency>
-
-    <dependency>
-      <groupId>de.ruedigermoeller</groupId>
-      <artifactId>fst</artifactId>
-      <version>2.56</version>
-    </dependency>
     <dependency>
       <groupId>com.vladsch.flexmark</groupId>
       <artifactId>flexmark-all</artifactId>