@dwightguth 2021-06-28
Given that this issue is affecting people who run `brew upgrade` on MacOS, what say we merge this early while we continue to work on improving performance?