@dwightguth 2021-06-28
Here is the performance data Radu asked for:

| Task | Time Before | Time After |
| ---- | -------------- | ----------- |
| Time to `kompile imp.md` full | 10.8s | 10.5s |
| Time to `kompile imp.md` incremental | 7.8s | 8.3s |
| Time to `kast collatz.imp` | 1.3s | 1.3s |
| Time to kompile KEVM full | 1m22s | 1m8s |
| Time to kompile KEVM incremental | 52.4s | 49.9s |
| Time to kast `add11.json` | 5.4s | 7s |

It looks like the time spent in kompile serializing and deserializing is not a major factor in the performance of kompile. The impact on performance in kast seems very small on small definitions, but is definitely somewhat noticable on larger ones. All in all, this isn't too bad, I don't think. Especially considering it doesn't impact the hot path.